Applying maxime recommendations
