+++ mgmtsystem_review/models/__init__.py
@@ -1,3 +1,4 @@
 # -*- coding: utf-8 -*-
 
+from. import mgmtsystem_review_line
 from . import mgmtsystem_review
