+++ mgmtsystem_review/report/review.xml
@@ -25,7 +25,7 @@
                     </div>
                     <div class="row" style="padding-top:10px">
                         <div class="col-xs-2">
-                            Confidentially:
+                            Confidentiality:
                         </div>
                         <div class="col-xs-10">
                             Direction Review
