+++ mgmtsystem_review/views/mgmtsystem_review.xml
@@ -14,7 +14,7 @@
                     <field name="action_id" attrs="{'invisible' : [('type', '!=', 'action')]}"/>
                     <field name="nonconformity_id" attrs="{'invisible' : [('type', '!=', 'nonconformity')]}"/>
                     <field name="decision"/>
-                    <field name="company_id" group="base.group_multi_company"/>
+                    <field name="company_id" groups="base.group_multi_company"/>
                 </tree>
             </field>
         </record>
@@ -30,7 +30,7 @@
                     <field name="action_id"/>
                     <field name="nonconformity_id"/>
                     <field name="decision"/>
-                    <field name="company_id" group="base.group_multi_company"/>
+                    <field name="company_id" groups="base.group_multi_company"/>
                 </search>
             </field>
         </record>
@@ -49,7 +49,7 @@
                         <field name="nonconformity_id" attrs="{'invisible' : [('type', '!=', 'nonconformity')]}"/>
                         <newline/>
                         <field name="decision"/>
-                        <field name="company_id" group="base.group_multi_company"/>
+                        <field name="company_id" groups="base.group_multi_company"/>
                     </group>
                 </form>
             </field>
@@ -73,7 +73,7 @@
                     <field name="name" attrs="{'readonly':[('state','=','done')]}"/>
                     <field name="date" attrs="{'readonly':[('state','=','done')]}"/>
                     <field name="reference"/>
-                    <field name="company_id" group="base.group_multi_company"/>
+                    <field name="company_id" groups="base.group_multi_company"/>
                     <!-- Invisible fields below are to allow their use in domains of this view -->
                     <field name="state" invisible="1"/>
                 </tree>
@@ -93,7 +93,7 @@
                         <field name="name"/>
                         <field name="date"/>
                         <field name="reference"/>
-                        <field name="company_id" group="base.group_multi_company"/>
+                        <field name="company_id" groups="base.group_multi_company"/>
                     </group>
                 </search>
             </field>
@@ -119,7 +119,7 @@
                             </group>
                             <group>
                                 <field name="reference"/>
-                                <field name="company_id" group="base.group_multi_company"/>
+                                <field name="company_id" groups="base.group_multi_company"/>
                             </group>
                         </group>
                         <notebook colspan="4">
@@ -135,7 +135,7 @@
                                 </group>
                                 <newline/>
                                 <separator string="Survey Answers" colspan="4"/>
-                                <field name="response_ids" colspan="4" nolabel="1" attrs="{'readonly':[('state','=','done')]}"/>
+                                <field name="response_ids" nolabel="1" attrs="{'readonly':[('state','=','done')]}"/>
                             </page>
                             <page string="Lines">
                                 <field name="line_ids" nolabel="1" attrs="{'readonly':[('state','=','done')]}"/>
@@ -172,7 +172,7 @@
             <field name="context">{"search_default_current":1}</field>
         </record>
 
-        <menuitem id="menu_open_review" action="open_mgmtsystem_review_list" parent="mgmtsystem.menu_mgmtsystem_main" sequence="0" groups="mgmtsystem.group_mgmtsystem_manager"/>
+        <menuitem id="menu_open_review" action="open_mgmtsystem_review_list" parent="mgmtsystem.menu_mgmtsystem_main" sequence="50" groups="mgmtsystem.group_mgmtsystem_manager"/>
 
     </data>
 </odoo>
